rename and merge
