--- drivers/bluetooth/sn32/iton/iton_driver.h
@@ -1,4 +1,0 @@
-#pragma once
-
-void iton_switch_driver(void);
-bool iton_is_bt(void);
