--- keyboards/gdmiller/gm807/config_led.h
@@ -1,48 +1,0 @@
-/*
-Copyright 2020 Adam Honse <calcprogrammer1@gmail.com>
-Copyright 2020 Dimitris Mantzouranis <d3xter93@gmail.com>
-Copyright 2022 1Conan <me@1conan.com>
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 2 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#pragma once
-
-#define LED_MATRIX_ROWS 6
-#define LED_MATRIX_ROW_CHANNELS 3
-#define LED_MATRIX_ROWS_HW (LED_MATRIX_ROWS * LED_MATRIX_ROW_CHANNELS)
-#define LED_MATRIX_ROW_PINS { C0, C1, C3, C4, C5, C6, C7, C8, C9, C10, C11, C12, C13, C14, B13, D3, B15, B14 }
-
-#define LED_MATRIX_COLS 17
-#define LED_MATRIX_COL_PINS { A8, A9, A10, A11, A12, A13, A14, A15, B0, B1, B2, B3, B4, B5, B6, B7, B8 }
-
-#define DRIVER_LED_TOTAL (87)
-
-#define RGB_MATRIX_MAXIMUM_BRIGHTNESS 255
-#define RGB_MATRIX_STARTUP_MODE RGB_MATRIX_SOLID_COLOR
-#define RGB_MATRIX_STARTUP_HUE 0
-#define RGB_MATRIX_STARTUP_SAT 0
-#define RGB_MATRIX_STARTUP_VAL 0
-#define RGB_MATRIX_STARTUP_SPD 0
-#define RGB_MATRIX_LED_FLUSH_LIMIT 32
-
-#define LED_CAPS_LOCK_PIN D6
-#define LED_K4 A6
-#define LED_K5 A7
-#define LED_USB_PORT D4
-
-// LEDs and Resistor unpopulated
-// Resistor pad is near the BT module
-// LEDs: LED6-10
-// Resistor: RA42 (Value: ? (probably an 82Ω))
-#define LED_ENCODER D5
