+++ keyboards/miller/gm807/gm807.h
@@ -1,4 +1,5 @@
 /* Copyright 2021 Harrison Chan (Xelus)
+ * Copyright 2022 1Conan <me@1conan.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -20,7 +21,7 @@
 
 #define XXX KC_NO
 
-#define LAYOUT_ansi( \
+#define LAYOUT( \
     k00, k01, k02, k03, k04, k05, k06, k07, k08, k09, k0a, k0b, k0c,      k0e, k0f, k0g, \
     k10, k11, k12, k13, k14, k15, k16, k17, k18, k19, k1a, k1b, k1c, k1d, k1e, k1f, k1g, \
     k20, k21, k22, k23, k24, k25, k26, k27, k28, k29, k2a, k2b, k2c, k2d, k2e, k2f, k2g, \
@@ -35,3 +36,12 @@
     { k40, XXX, k42, k43, k44, k45, k46, k47, k48, k49, k4a, k4b, XXX, k4d, XXX, k4f, XXX }, \
     { k50, k51, k52, XXX, XXX, XXX, k56, XXX, XXX, XXX, k5a, k5b, k5c, k5d, k5e, k5f, k5g }  \
 }
+
+enum GM807_keycodes {
+    KEY_USB_LED = SAFE_RANGE,
+    BT_PAIR,
+    BT_PRO1,
+    BT_PRO2,
+    BT_PRO3,
+    GM807_SAFE_RANGE,
+};
