+++ keyboards/miller/gm807/info.json
@@ -1,7 +1,7 @@
 {
-    "keyboard_name": "Rakk Lam Ang Pro RGB",
+    "keyboard_name": "miller GM807",
     "url": "",
-    "maintainer": "xelus22",
+    "maintainer": "1Conan",
     "width": 17,
     "height": 6,
     "layouts": {
