+++ keyboards/miller/gm807/readme.md
@@ -1,13 +1,19 @@
-# gdmiller GM807
+# miller GM807
 
-GM807 Hotswap RGB ANSI
+miller GM807 Hotswap RGB
+
+Also sold as:
+
+- Akko K870T
+- Flashquark Falcon Elite
+- Keebio Hotswap TKL Bluetooth Keyboard w/RGB
+- Rakk Lam-Ang (Pro)
 
 Make example for this keyboard (after setting up your build environment):
-```
-make gdmiller/gm807:default
-```
+
+    make miller/gm807:default
 
 * * *
-For more info and detailed flashing instructions, please visit https://github.com/1Conan/gdmiller-GM807
+For more info and detailed flashing instructions, please visit https://github.com/1Conan/miller-gm807
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
