--- keyboards/rakk_lam_ang_pro/readme.md
@@ -1,19 +1,0 @@
-# Rakk Lam Ang Pro
-
-Rakk Lam Ang Pro Hotswap RGB
-
-Make example for this keyboard (after setting up your build environment):
-
-    make rakk_lam_ang_pro:default
-
-* * *
-For more info and detailed flashing instructions, please visit https://github.com/alex-dinh/keychron-k8
-
-Reset your keyboard in 3 ways:
-<ol>
-<li>Software reset on Fn + Backspace</li>
-<li>Bootmagic reset: hold down the top left key (usually escape) and plugin the keyboard</li>
-<li>Physical reset: Short Boot0 while plugging in the keyboard</li>
-</ol>
-
-See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
